rm cloud run job
